requirements.txt
measurement>=1.6
django-autocomplete-light>=3.1.5
